+++ src/Form.jsx
@@ -1,51 +1,34 @@
+import RestaurantInput from './RestaurantInput';
+
 export default function Form({
   value,
   onChange,
   onClick,
 }) {
-  function handleChange(event) {
-    const { name, value: changedValue } = event.target;
-
-    if (name === 'name') {
-      onChange({ restaurantName: changedValue });
-    }
-
-    if (name === 'category') {
-      onChange({ restaurantCategory: changedValue });
-    }
-
-    if (name === 'address') {
-      onChange({ restaurantAddress: changedValue });
-    }
-  }
-
   const { name, category, address } = value;
 
   return (
     <p>
-      <input
+      <RestaurantInput
         id="input-restaurant-name"
         name="name"
-        type="text"
         placeholder="이름"
         value={name}
-        onChange={handleChange}
+        onChange={onChange}
       />
-      <input
+      <RestaurantInput
         id="input-restaurant-category"
         name="category"
-        type="text"
         placeholder="분류"
         value={category}
-        onChange={handleChange}
+        onChange={onChange}
       />
-      <input
+      <RestaurantInput
         id="input-restaurant-address"
         name="address"
-        type="text"
         placeholder="주소"
         value={address}
-        onChange={handleChange}
+        onChange={onChange}
       />
       <button type="button" onClick={onClick}>
         등록
